requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
pydantic_settings
sqlalchemy
psycopg2-binary
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client